refactor: match filenames with the class and improve others to clarify
